Screenshots of functionality
Uploading four screenshots. Two screenshots are showing the coordinates (which are located at the top left of the screen in format XYZ: X / Y / Z) of the two corners used in the test, one screenshot is showing an overview of the structure being analyzed, and the other one shows the python program running after all inputs are process.